requirements_gui.txt
# GUI Application Requirements
numpy>=1.19.0
pyinstaller>=4.0
pillow>=8.0.0

